This is a website for my final year project.
A dog snoozing on a street corner is not an unusual sight in Sri Lanka. There are many street dogs that are homeless and live on the streets. Most street dogs are unhealthy due to a lack of attention for their well-being. Some of the street dogs are diagnosed with cancer, skin rashes, and rabies. So, there are several cases where people are bitten by street dogs and there are only a few cases where people die because of rabies. To reduce the street dog population many countries including Sri Lanka use cruel methods like killing by injection. Many Animal Welfare Organizations (AWOs) had been up for the defense of this act. Luckily there are communities willing to save street dogs out of their misery and give them a comfortable living during the short span of years they live. Dogee is one of the major organizations in Sri Lanka that helps street dogs.